+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 * Fix pointer scan for 32-bit games
 * Fix audio mixing with non-integer samples per frame (#600)
 * Inserting a marker does update correctly the marker table
+* Add sanity check when no buffer size is provided in SDL_OpenAudio()
 
 ## [1.4.5] - 2023-10-22
 ### Added